Part 2/deployment/requirements.txt
streamlit
pandas
matplotlib
numpy
scikit-learn==1.0.2
tensorflow 
Pillow
openpyxl
opencv-python
